refactor: rename IPost to Post
