+++ src/components/pages/user-page.tsx
@@ -11,7 +11,7 @@ type Props = {
   isProfileOfLoggerUser: boolean;
 };
 
-function UserPage(props: Props) {
+export function UserDetailPage(props: Props) {
   return (
     <div className="flex flex-col h-full max-xl:border-0 gap-2 pl-3 pr-3 border-rebeccapurple2 border-r-2 border-l-2">
       <Header title={props.profile.username} />
@@ -26,4 +26,3 @@ function UserPage(props: Props) {
     </div>
   );
 };
-export default UserPage;
